chore(tests): isolate concurrency related tests in their own class
