+++ jdbc-h2/src/test/java/io/kestra/runner/h2/H2RunnerConcurrencyTest.java
@@ -1,0 +1,6 @@
+package io.kestra.runner.h2;
+
+import io.kestra.core.runners.AbstractRunnerConcurrencyTest;
+
+public class H2RunnerConcurrencyTest extends AbstractRunnerConcurrencyTest {
+}
