+++ jdbc-mysql/src/test/java/io/kestra/runner/mysql/MysqlRunnerConcurrencyTest.java
@@ -1,0 +1,6 @@
+package io.kestra.runner.mysql;
+
+import io.kestra.core.runners.AbstractRunnerConcurrencyTest;
+
+public class MysqlRunnerConcurrencyTest extends AbstractRunnerConcurrencyTest {
+}
