+++ jdbc-postgres/src/test/java/io/kestra/runner/postgres/PostgresRunnerConcurrencyTest.java
@@ -1,0 +1,6 @@
+package io.kestra.runner.postgres;
+
+import io.kestra.core.runners.AbstractRunnerConcurrencyTest;
+
+public class PostgresRunnerConcurrencyTest extends AbstractRunnerConcurrencyTest {
+}
